requirements.txt
redis
flask
ecdsa
pynacl
